+++ headless_kivy_pi/constants.py
@@ -12,12 +12,15 @@ BITS_PER_BYTE = 11
 
 # Configure the headless mode for the Kivy application and initialize the display
 
+IS_TEST_ENVIRONMENT = (
+    strtobool(os.environ.get('HEADLESS_KIVY_PI_TEST_ENVIRONMENT', 'False')) == 1
+)
 MIN_FPS = int(os.environ.get('HEADLESS_KIVY_PI_MIN_FPS', '1'))
 MAX_FPS = int(os.environ.get('HEADLESS_KIVY_PI_MAX_FPS', '32'))
 WIDTH = int(os.environ.get('HEADLESS_KIVY_PI_WIDTH', '240'))
 HEIGHT = int(os.environ.get('HEADLESS_KIVY_PI_HEIGHT', '240'))
 BAUDRATE = int(os.environ.get('HEADLESS_KIVY_PI_BAUDRATE', '60000000'))
-DEBUG_MODE = (
+IS_DEBUG_MODE = (
     strtobool(
         os.environ.get('HEADLESS_KIVY_PI_DEBUG', 'False' if IS_RPI else 'True'),
     )
@@ -35,9 +38,9 @@ SYNCHRONOUS_CLOCK = (
     )
     == 1
 )
-AUTOMATIC_FPS_CONTROL = (
+AUTOMATIC_FPS = (
     strtobool(
-        os.environ.get('HEADLESS_KIVY_PI_AUTOMATIC_FPS_CONTROL', 'True'),
+        os.environ.get('HEADLESS_KIVY_PI_AUTOMATIC_FPS', 'True'),
     )
     == 1
 )